feat: #22-HMAC을 이용한 검증
