+++ back/src/main/java/com/example/capstone/global/dto/HmacRequest.java
@@ -1,0 +1,2 @@
+package com.example.capstone.global.dto;public interface HmacRequest {
+}
